+++ src/main/java/dinkplugin/message/NotificationBody.java
@@ -1,7 +1,9 @@
 package dinkplugin.message;
 
 import com.google.gson.annotations.SerializedName;
+import dinkplugin.DinkPluginConfig;
 import dinkplugin.notifiers.data.NotificationData;
+import dinkplugin.util.DiscordProfile;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -36,6 +38,15 @@ public class NotificationBody<T extends NotificationData> {
     /*
      * Discord fields
      */
+
+    /**
+     * Information about the current discord user, acquired via RPC (handled by base RuneLite).
+     * <p>
+     * This is only sent if {@link DinkPluginConfig#sendDiscordUser()} is enabled.
+     * While this field is not used by Discord, it can be useful for custom webhook handlers that forward to Discord.
+     */
+    @Nullable
+    DiscordProfile discordUser;
 
     /**
      * Filled in with the text of the notifier (e.g., {@link #getText()} is "Forsen has levelled Attack to 100")