+++ src/test/java/dinkplugin/notifiers/MockedNotifierTest.java
@@ -20,6 +20,7 @@ import net.runelite.client.callback.ClientThread;
 import net.runelite.client.chat.ChatMessageManager;
 import net.runelite.client.config.ConfigDescriptor;
 import net.runelite.client.config.ConfigManager;
+import net.runelite.client.discord.DiscordService;
 import net.runelite.client.game.ItemManager;
 import net.runelite.client.game.NPCManager;
 import net.runelite.client.ui.DrawManager;
@@ -69,6 +70,9 @@ abstract class MockedNotifierTest extends MockedTestBase {
     protected ScheduledExecutorService executor = new BlockingExecutor();
 
     @Bind
+    protected DiscordService discordService = Mockito.mock(DiscordService.class);
+
+    @Bind
     protected ChatMessageManager chatManager = Mockito.mock(ChatMessageManager.class);
 
     @Bind
@@ -87,7 +91,7 @@ abstract class MockedNotifierTest extends MockedTestBase {
     protected SettingsManager settingsManager = Mockito.spy(new SettingsManager(gson, client, clientThread, plugin, config, configManager));
 
     @Bind
-    protected DiscordMessageHandler messageHandler = Mockito.spy(new DiscordMessageHandler(gson, client, drawManager, httpClient, config, executor, clientThread));
+    protected DiscordMessageHandler messageHandler = Mockito.spy(new DiscordMessageHandler(gson, client, drawManager, httpClient, config, executor, clientThread, discordService));
 
     @Override
     protected void setUp() {